all works) todo upload avatar
