+++ src/Store/monthStat/monthStatThunk.js
@@ -1,10 +1,12 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { fetchMonthStat } from 'services/api';
+import { fetchMonthStat } from 'services/api.js';
 
-export const getMonthStat = createAsyncThunk('monthStat/getMonthStat', async ({month, year}, thunkApi) => {
+export const getMonthStat = createAsyncThunk(
+  'monthStat/getMonthStat',
+  async ({ month, year }, thunkApi) => {
     try {
-        const monthStat = await fetchMonthStat(month, year);
-        return monthStat.data;
+      const monthStat = await fetchMonthStat(month, year);
+      return monthStat.data;
     } catch (error) {
       return thunkApi.rejectWithValue(error.message);
     }
